fix(flashcards): split legacy-compat to keep Prisma out of client bundle
Sprint 6's legacy-compat.ts had a Prisma value import (for
resolveLegacyDeckId) AND was re-exported through the flashcards
barrel. Client components like FlashcardReviewOverlay import from
that barrel, which transitively pulled pg into the browser bundle and
broke the Turbopack build with:

  Module not found: Can't resolve 'dns'  (pg/lib/connection-parameters.js)
  Module not found: Can't resolve 'tls'  (pg/lib/stream.js)
  Module not found: Can't resolve 'net'  (pg/lib/stream.js)
  Module not found: Can't resolve 'fs'   (pgpass, pg-connection-string)
  + chunking context doesn't support node:async_hooks / node:fs

CLAUDE.md explicitly warns about this:
  "NEVER import Prisma into 'use client' components"

The pattern documented in the project for this exact problem is the
ai/tools split:
  - lib/domain/ai/tools/metadata.ts — client-safe, no Prisma
  - lib/domain/ai/tools/registry.ts — server-only, has Prisma

Applying the same split to the flashcards legacy-compat surface:

  lib/domain/flashcards/legacy-derive.ts (NEW)
    Pure functions, type-only Prisma imports. Re-exported through the
    barrel. Contains:
      - deriveLegacyCategoryAndSubcategory(deck)
      - deriveLegacyReviewStatus(state, reps, lapses)
      - deriveStateFromLegacyStatus(status, now)
      - LegacyDeckSource type

  lib/domain/flashcards/legacy-compat.ts
    Server-only. Has the Prisma value import. NOT re-exported through
    the barrel. Server callers import via deep path:
      import { resolveLegacyDeckId } from "@/lib/domain/flashcards/legacy-compat"
    Re-exports the legacy-derive surface so server callers have a
    single import path.

Updated:
  - lib/domain/flashcards/index.ts — re-exports legacy-derive only,
    with an inline comment explaining why
  - lib/domain/flashcards/api.ts — imports derive helpers from
    legacy-derive (the file it depends on doesn't transitively pull
    in Prisma)
  - app/api/flashcards/route.ts — deep import for resolveLegacyDeckId
  - app/api/flashcards/[id]/route.ts — deep import
  - app/api/flashcards/decks/route.ts — deep import

The barrel still exports the derive helpers (deriveLegacyCategoryAnd-
Subcategory, deriveLegacyReviewStatus, deriveStateFromLegacyStatus),
so any caller that only needed those still imports them via
@/lib/domain/flashcards without a code change.

Verified
  - tsc --noEmit exit 0
